6483-Step1: move validation code out of pipeline_types.go
As part of the step 1 of https://github.com/tektoncd/pipeline/issues/6483,
this commit moves validation function from pipeline_types.go to pipeline_validation.go

No impact on the downstream consumers b/c they stay in the same package.
